+++ NEWS.md
@@ -1,3 +1,5 @@
+# SpaDES.tools (development version)
+
 # SpaDES.tools 2.1.3
 
 ## Deprecated and defunct